requirements.txt
amqp==5.1.1
appnope==0.1.3
APScheduler==3.9.1
asgiref==3.5.1
asttokens==2.0.5
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
bleach==5.0.0
cachetools==5.2.0
celery==5.2.7
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Deprecated==1.2.13
Django==3.2.13
django-celery-beat==2.3.0
django-celery-results==2.3.1
django-cors-headers==3.13.0
django-crontab==0.7.1
django-extensions==3.2.1
django-filter==21.1
django-http-referrer-policy==1.1.1
django-otp==1.1.3
django-redis==5.2.0
django-referrer-policy==1.0
django-rest-framework==0.1.0
django-sslserver==0.22
django-summernote==0.8.20.0
django-timezone-field==5.0
djangorestframework==3.13.1
drf-yasg==1.20.0
executing==0.8.3
fanstatic==1.2
google-api-core==2.8.2
google-auth==2.11.0
google-cloud-aiplatform==1.16.1
google-cloud-bigquery==2.34.4
google-cloud-core==2.3.2
google-cloud-resource-manager==1.6.1
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
googletrans==3.0.0
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-status==1.47.0
h11==0.9.0
h2==3.2.0
haversine==2.7.0
hpack==3.0.0
hstspreload==2022.10.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib==1.0.4
importlib-metadata==4.11.3
inflection==0.5.1
ipdb==0.13.9
ipython==8.4.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
js.jquery==3.3.1
js.jqueryui==1.10.3
kombu==5.2.4
libretranslatepy==2.1.1
lxml==4.9.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
numpy==1.22.4
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
proto-plus==1.22.0
protobuf==3.20.1
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.1
requests==2.27.1
rfc3986==1.5.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
shutilwhich==1.1.0
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
traitlets==5.2.2.post1
translate==3.6.1
twilio==7.14.1
tzdata==2022.2
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
